+++ lang/include/Byte.hpp
@@ -12,7 +12,7 @@ class _String;
 #define byte uint8_t
 
 //define uint16_t to avoid sstream change assic to character.
-DECLARE_CLASS(Byte) IMPLEMENTS(Number)<uint16_t> {
+DECLARE_CLASS(Byte) IMPLEMENTS(Number)<uint8_t> {
 public:
     static const int __isReflected = 1;
